refactor: rename from shisen_sim into shisen_gazebo_plugins
