+++ README.md
@@ -1,4 +1,4 @@
-# Shisen Sim
+# Shisen Gazebo Plugins
 
 [![latest version](https://img.shields.io/github/v/release/threeal/shisen_sim.svg)](https://github.com/threeal/shisen_sim/releases/)
 [![license](https://img.shields.io/github/license/threeal/shisen_sim.svg)](./LICENSE)
