+++ package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>shisen_sim</name>
+  <name>shisen_gazebo_plugins</name>
   <version>0.1.0</version>
   <description>A Gazebo simulation version of Shisen movement package</description>
   <maintainer email="alfi.maulana.f@gmail.com">Alfi Maulana</maintainer>
